backend/requirements.txt
boto3
stripe
psycopg2-binary
